docs: adding docs to generate the correct UI flow
